Move the report view into a new `repo show` command
`repo search` prints the match table alone, one output shape whatever
the cache holds. `repo show` takes the same query and prints every
matched identity's full report in match order, as `apt show` prints
one record per named package; a query nothing matches is an error, and
so is a matched identity one repository publishes twice. The
settled-identity heuristic and the `--full` flag are removed with the
split, along with the JSON `detail`/`details` fields: search JSON
carries `query`/`matches`/`excluded`, show JSON adds `reports`.
