+++ crates/peppy/tests/repo_search.rs
@@ -2,7 +2,7 @@ use super::common::{ARM_LINK_PAIRING, publish_repo_index, setup};
 use config::consts::NODE_CONFIG_FILE;
 use daemon_config::consts::PeppyDirs;
 use peppy::commands::Command;
-use peppy::commands::repo::{RepoCommand, RepoCommands, search_rendered};
+use peppy::commands::repo::{RepoCommand, RepoCommands, search_rendered, show_rendered};
 use std::path::{Path, PathBuf};
 
 const RGB_CAMERA_CONTRACT: &str = r#"{
@@ -140,16 +140,16 @@ fn seeded(
     (peppy_dirs, repo_dir, contract_sha)
 }
 
-/// A refreshed repository answers a contract search from the caches: the
+/// A refreshed repository answers a contract show from the caches: the
 /// published document, the implementer with its pin checked, and the
 /// consumer with its cardinality, each at the manifest path refresh
 /// recorded.
 #[test]
-fn repo_search_reports_a_contracts_implementers_and_consumers() {
+fn repo_show_reports_a_contracts_implementers_and_consumers() {
     let (_rt, serve, ctx, _work_dir) = setup();
     let (peppy_dirs, repo_dir, contract_sha) = seeded(&serve, &ctx);
 
-    let text = search_rendered(&peppy_dirs, "rgb_camera:v1", false, false, None).expect("searches");
+    let text = show_rendered(&peppy_dirs, "rgb_camera:v1", false, None).expect("shows");
 
     let label = repo_dir.to_string_lossy();
     assert!(
@@ -185,14 +185,14 @@ fn repo_search_reports_a_contracts_implementers_and_consumers() {
     );
 }
 
-/// A pairing search reports the roles played and observed, with the role
+/// A pairing show reports the roles played and observed, with the role
 /// and slot facts the manifests declare.
 #[test]
-fn repo_search_reports_a_pairings_participants_and_observers() {
+fn repo_show_reports_a_pairings_participants_and_observers() {
     let (_rt, serve, ctx, _work_dir) = setup();
     let (peppy_dirs, repo_dir, _) = seeded(&serve, &ctx);
 
-    let text = search_rendered(&peppy_dirs, "arm_link:v1", false, false, None).expect("searches");
+    let text = show_rendered(&peppy_dirs, "arm_link:v1", false, None).expect("shows");
 
     assert!(
         text.contains(&format!(
@@ -224,11 +224,11 @@ fn repo_search_reports_a_pairings_participants_and_observers() {
 /// A bare name is a pattern matching any tag: the contract's report comes
 /// back without the query naming `v1`.
 #[test]
-fn repo_search_a_bare_name_matches_any_tag() {
+fn repo_show_a_bare_name_matches_any_tag() {
     let (_rt, serve, ctx, _work_dir) = setup();
     let (peppy_dirs, repo_dir, contract_sha) = seeded(&serve, &ctx);
 
-    let text = search_rendered(&peppy_dirs, "rgb_camera", false, false, None).expect("searches");
+    let text = show_rendered(&peppy_dirs, "rgb_camera", false, None).expect("shows");
 
     assert!(
         text.starts_with(&format!(
@@ -248,7 +248,7 @@ fn repo_search_lists_partial_matches() {
     let (_rt, serve, ctx, _work_dir) = setup();
     let (peppy_dirs, repo_dir, contract_sha) = seeded(&serve, &ctx);
 
-    let text = search_rendered(&peppy_dirs, "camera", false, false, None).expect("searches");
+    let text = search_rendered(&peppy_dirs, "camera", false, None).expect("searches");
 
     assert!(
         text.starts_with("camera\n\n3 items match `camera`\n"),
@@ -283,14 +283,14 @@ fn repo_search_lists_partial_matches() {
     );
 }
 
-/// `--full` prints a report block for every matched identity instead of
-/// the table, in match order.
+/// `repo show` prints a report block for every matched identity instead
+/// of the table, in match order.
 #[test]
-fn repo_search_full_reports_every_match() {
+fn repo_show_reports_every_match() {
     let (_rt, serve, ctx, _work_dir) = setup();
     let (peppy_dirs, _repo_dir, _) = seeded(&serve, &ctx);
 
-    let text = search_rendered(&peppy_dirs, "camera", true, false, None).expect("searches");
+    let text = show_rendered(&peppy_dirs, "camera", false, None).expect("shows");
 
     assert!(!text.contains("items match"), "{text}");
     let exposure = text
@@ -310,14 +310,19 @@ fn repo_search_full_reports_every_match() {
     );
 }
 
-/// A launcher is found by its file stem, the one untagged kind: its
-/// published line is the whole answer.
+/// A launcher is found by its file stem, the one untagged kind: the
+/// search lists its one row in the singular, and its published line is
+/// the show's whole answer.
 #[test]
-fn repo_search_finds_a_launcher() {
+fn repo_show_details_a_launcher() {
     let (_rt, serve, ctx, _work_dir) = setup();
     let (peppy_dirs, repo_dir, _) = seeded(&serve, &ctx);
 
-    let text = search_rendered(&peppy_dirs, "demo", false, false, None).expect("searches");
+    let listed = search_rendered(&peppy_dirs, "demo", false, None).expect("searches");
+    assert!(listed.contains("\n1 item matches `demo`\n"), "{listed}");
+    assert!(listed.contains("│ launcher │ demo │"), "{listed}");
+
+    let text = show_rendered(&peppy_dirs, "demo", false, None).expect("shows");
 
     assert_eq!(
         text,
@@ -332,14 +337,14 @@ fn repo_search_finds_a_launcher() {
 
 /// A digest query keeps only the copies carrying those bytes: the
 /// published fingerprint finds the document, an unpublished one finds
-/// nothing.
+/// nothing, which a search answers and a show refuses.
 #[test]
-fn repo_search_finds_a_pinned_digest() {
+fn repo_show_finds_a_pinned_digest() {
     let (_rt, serve, ctx, _work_dir) = setup();
     let (peppy_dirs, repo_dir, contract_sha) = seeded(&serve, &ctx);
 
     let pinned = format!("rgb_camera:v1@{contract_sha}");
-    let text = search_rendered(&peppy_dirs, &pinned, false, false, None).expect("searches");
+    let text = show_rendered(&peppy_dirs, &pinned, false, None).expect("shows");
     assert!(
         text.starts_with(&format!(
             "{pinned}\n  contract rgb_camera:v1 published by {} at {} (sha256 {contract_sha})\n",
@@ -351,28 +356,30 @@ fn repo_search_finds_a_pinned_digest() {
     assert!(text.contains("\nImplemented by 1 indexed node\n"), "{text}");
 
     let absent = "f".repeat(64);
-    let text = search_rendered(
-        &peppy_dirs,
-        &format!("rgb_camera:v1@{absent}"),
-        false,
-        false,
-        None,
-    )
-    .expect("an empty answer");
+    let text = search_rendered(&peppy_dirs, &format!("rgb_camera:v1@{absent}"), false, None)
+        .expect("an empty answer");
     assert!(
         text.contains("nothing in any configured repository's cache matches"),
         "{text}"
     );
+
+    let err = show_rendered(&peppy_dirs, &format!("rgb_camera:v1@{absent}"), false, None)
+        .expect_err("a show has no report to print")
+        .to_string();
+    assert!(
+        err.contains("nothing in any configured repository's cache matches"),
+        "{err}"
+    );
 }
 
-/// The JSON form of the same search parses and carries the matches and
-/// the report.
+/// The JSON form of a show parses and carries the matches and the
+/// report.
 #[test]
-fn repo_search_json_carries_the_report() {
+fn repo_show_json_carries_the_report() {
     let (_rt, serve, ctx, _work_dir) = setup();
     let (peppy_dirs, _repo_dir, contract_sha) = seeded(&serve, &ctx);
 
-    let text = search_rendered(&peppy_dirs, "rgb_camera:v1", false, true, None).expect("searches");
+    let text = show_rendered(&peppy_dirs, "rgb_camera:v1", true, None).expect("shows");
     let doc: serde_json::Value = serde_json::from_str(&text).expect("valid JSON");
 
     assert_eq!(doc["query"]["raw"], "rgb_camera:v1");
@@ -385,40 +392,43 @@ fn repo_search_json_carries_the_report() {
     assert_eq!(matches[0]["exact"], true);
     assert_eq!(matches[0]["published"]["sha256"], contract_sha);
     assert_eq!(matches[0]["published"]["repo_id"], 1);
-    let detail = &doc["detail"];
-    assert_eq!(detail["name"], "rgb_camera");
-    assert_eq!(detail["implementers"][0]["node"]["node_name"], "uvc_camera");
-    assert_eq!(detail["implementers"][0]["node"]["source_type"], "fs");
-    assert_eq!(detail["implementers"][0]["sha256"], contract_sha);
-    assert_eq!(detail["implementers"][0]["pin"]["status"], "current");
-    assert_eq!(detail["consumers"][0]["node"]["node_name"], "recorder");
-    assert_eq!(detail["consumers"][0]["cardinality"], "zero_or_one");
-    assert_eq!(detail["participants"], serde_json::json!([]));
-    assert_eq!(detail["observers"], serde_json::json!([]));
+    assert_eq!(doc["reports"].as_array().expect("array").len(), 1);
+    let report = &doc["reports"][0];
+    assert_eq!(report["name"], "rgb_camera");
+    assert_eq!(report["implementers"][0]["node"]["node_name"], "uvc_camera");
+    assert_eq!(report["implementers"][0]["node"]["source_type"], "fs");
+    assert_eq!(report["implementers"][0]["sha256"], contract_sha);
+    assert_eq!(report["implementers"][0]["pin"]["status"], "current");
+    assert_eq!(report["consumers"][0]["node"]["node_name"], "recorder");
+    assert_eq!(report["consumers"][0]["cardinality"], "zero_or_one");
+    assert_eq!(report["participants"], serde_json::json!([]));
+    assert_eq!(report["observers"], serde_json::json!([]));
 
-    let text = search_rendered(&peppy_dirs, "arm_link:v1", false, true, None).expect("searches");
+    let text = show_rendered(&peppy_dirs, "arm_link:v1", true, None).expect("shows");
     let doc: serde_json::Value = serde_json::from_str(&text).expect("valid JSON");
     assert_eq!(doc["matches"][0]["kind"], "pairing");
-    let detail = &doc["detail"];
-    assert_eq!(detail["participants"][0]["role"], "arm");
-    assert_eq!(detail["participants"][0]["optional"], true);
-    assert_eq!(detail["observers"][0]["role"], "controller");
-    assert_eq!(detail["observers"][0]["cardinality"], "one");
+    let report = &doc["reports"][0];
+    assert_eq!(report["participants"][0]["role"], "arm");
+    assert_eq!(report["participants"][0]["optional"], true);
+    assert_eq!(report["observers"][0]["role"], "controller");
+    assert_eq!(report["observers"][0]["cardinality"], "one");
 }
 
-/// The JSON of several matches lists each kind, and the launcher's
-/// missing tag is `null`. The pattern spells out only `demo` in full, so
-/// the query settles on the launcher and its (empty) report rides along.
+/// The JSON of a search lists each kind with no usage report, and the
+/// launcher's missing tag is `null`.
 #[test]
 fn repo_search_json_lists_matches() {
     let (_rt, serve, ctx, _work_dir) = setup();
     let (peppy_dirs, _repo_dir, _) = seeded(&serve, &ctx);
 
-    let text = search_rendered(&peppy_dirs, "camera|demo", false, true, None).expect("searches");
+    let text = search_rendered(&peppy_dirs, "camera|demo", true, None).expect("searches");
     let doc: serde_json::Value = serde_json::from_str(&text).expect("valid JSON");
 
-    assert_eq!(doc["detail"]["name"], "demo");
-    assert_eq!(doc["detail"]["implementers"], serde_json::json!([]));
+    assert_eq!(
+        doc.as_object().expect("object").len(),
+        3,
+        "a search carries `query`, `matches`, and `excluded`, nothing more: {doc}"
+    );
     let kinds: Vec<&str> = doc["matches"]
         .as_array()
         .expect("array")
@@ -443,7 +453,7 @@ fn repo_search_rejects_an_invalid_regex() {
     let (_rt, serve, _ctx, _work_dir) = setup();
     let peppy_dirs = PeppyDirs::new(serve.temp_dir());
 
-    let err = search_rendered(&peppy_dirs, "rgb_camera[", false, false, None)
+    let err = search_rendered(&peppy_dirs, "rgb_camera[", false, None)
         .expect_err("an unclosed class is not a pattern")
         .to_string();
 
@@ -460,7 +470,7 @@ fn repo_search_rejects_a_bad_digest() {
     let (_rt, serve, _ctx, _work_dir) = setup();
     let peppy_dirs = PeppyDirs::new(serve.temp_dir());
 
-    let err = search_rendered(&peppy_dirs, "rgb_camera@xyz", false, false, None)
+    let err = search_rendered(&peppy_dirs, "rgb_camera@xyz", false, None)
         .expect_err("`xyz` is not a fingerprint")
         .to_string();
 
@@ -474,8 +484,7 @@ fn repo_search_says_when_nothing_matches() {
     let (_rt, serve, ctx, _work_dir) = setup();
     let (peppy_dirs, _repo_dir, _) = seeded(&serve, &ctx);
 
-    let text =
-        search_rendered(&peppy_dirs, "nobody:v1", false, false, None).expect("an empty answer");
+    let text = search_rendered(&peppy_dirs, "nobody:v1", false, None).expect("an empty answer");
 
     assert_eq!(
         text,
@@ -489,13 +498,12 @@ fn repo_search_says_when_nothing_matches() {
 /// outline never exceeds it, and the over-long absolute manifest paths
 /// wrap onto continuation lines instead of breaking the box.
 #[test]
-fn repo_search_wraps_tables_to_a_narrow_width() {
+fn repo_show_wraps_tables_to_a_narrow_width() {
     let (_rt, serve, ctx, _work_dir) = setup();
     let (peppy_dirs, _repo_dir, _) = seeded(&serve, &ctx);
 
-    let wide = search_rendered(&peppy_dirs, "rgb_camera:v1", false, false, None).expect("searches");
-    let narrow =
-        search_rendered(&peppy_dirs, "rgb_camera:v1", false, false, Some(60)).expect("searches");
+    let wide = show_rendered(&peppy_dirs, "rgb_camera:v1", false, None).expect("shows");
+    let narrow = show_rendered(&peppy_dirs, "rgb_camera:v1", false, Some(60)).expect("shows");
 
     for line in narrow.lines() {
         if ['│', '┐', '┤', '┘'].iter().any(|c| line.ends_with(*c)) {
