+++ docs/src/content/docs/advanced_guides/contract_implementation.mdx
@@ -54,7 +54,7 @@ A contract lives in its own file under a repository Peppy scans (see [Repositori
 }
 ```
 
-After `peppy repo refresh`, the contract is cached and addressable by `(name, tag)`. Files use whatever filename you like as long as it ends in `.json5`; Peppy identifies a `contract/v1` document by its `peppy_schema` field. `peppy repo search <name>:<tag>` lists the indexed nodes that implement or consume it (see [Search the repositories](/advanced_guides/repositories/#search-the-repositories)).
+After `peppy repo refresh`, the contract is cached and addressable by `(name, tag)`. Files use whatever filename you like as long as it ends in `.json5`; Peppy identifies a `contract/v1` document by its `peppy_schema` field. `peppy repo show <name>:<tag>` lists the indexed nodes that implement or consume it (see [Show what a query matches](/advanced_guides/repositories/#show-what-a-query-matches)).
 
 ### 2. A producer that implements the contract
 
@@ -525,7 +525,7 @@ manifest: {
 }
 ```
 
-Each side independently verifies its pinned `sha256` against the on-disk contract document at cache-resolution time. Peppy refuses to start a node whose pinned contract revision is not in the cache. The two sides are not cross-checked: a producer that pins `sha256` against the cached contract and a consumer that does not pin can still bind, as long as both pass their own checks. `peppy repo search <name>:<tag>` shows every implementer's and consumer's pin next to the published document's fingerprint, so a pin a sync would refuse is visible before the sync.
+Each side independently verifies its pinned `sha256` against the on-disk contract document at cache-resolution time. Peppy refuses to start a node whose pinned contract revision is not in the cache. The two sides are not cross-checked: a producer that pins `sha256` against the cached contract and a consumer that does not pin can still bind, as long as both pass their own checks. `peppy repo show <name>:<tag>` shows every implementer's and consumer's pin next to the published document's fingerprint, so a pin a sync would refuse is visible before the sync.
 
 ## Worked example: multi-camera reconstruction
