+++ docs/src/content/docs/advanced_guides/pairing.mdx
@@ -60,7 +60,7 @@ The rules:
 - `topics` is one flat list; every topic's `emitted_by` must name one of the two roles, and topic names must be unique across the whole list.
 - Pairings declare **topics only**: no services, no actions.
 
-After `peppy repo refresh`, the pairing is cached and addressable by `(name, tag)`. `peppy repo search <name>:<tag>` lists the indexed nodes that play its roles and the ones that observe them (see [Search the repositories](/advanced_guides/repositories/#search-the-repositories)).
+After `peppy repo refresh`, the pairing is cached and addressable by `(name, tag)`. `peppy repo show <name>:<tag>` lists the indexed nodes that play its roles and the ones that observe them (see [Show what a query matches](/advanced_guides/repositories/#show-what-a-query-matches)).
 
 ## Configure the nodes
